Move clientfactory from pkg/syncer/syncers to pkg/syncer
