+++ pkg/syncer/syncers/common.go
@@ -26,6 +26,7 @@ import (
 	"k8s.io/klog/v2"
 
 	edgev1alpha1 "github.com/kcp-dev/edge-mc/pkg/apis/edge/v1alpha1"
+	. "github.com/kcp-dev/edge-mc/pkg/syncer/clientfactory"
 )
 
 func resourceToString(resource edgev1alpha1.EdgeSyncConfigResource) string {
