+++ pkg/syncer/syncers/downsyncer.go
@@ -27,6 +27,7 @@ import (
 	"k8s.io/klog/v2"
 
 	edgev1alpha1 "github.com/kcp-dev/edge-mc/pkg/apis/edge/v1alpha1"
+	. "github.com/kcp-dev/edge-mc/pkg/syncer/clientfactory"
 )
 
 type DownSyncer struct {
