+++ pkg/syncer/syncers/upsyncer.go
@@ -26,6 +26,7 @@ import (
 	"k8s.io/klog/v2"
 
 	edgev1alpha1 "github.com/kcp-dev/edge-mc/pkg/apis/edge/v1alpha1"
+	. "github.com/kcp-dev/edge-mc/pkg/syncer/clientfactory"
 )
 
 type UpSyncer struct {
